Add support for a post-tonemap color LUT
This is exposed as "sdr_lut" since it uses an standard dynamic range
(SDR) input and SDR output. This means this operation is intended for
SDR displays/outputs.

A LUT texture must be generated to use this feature. To do this, first
save a screenshot with an embedded LUT texture using
`simplepbr.sdr_lut_sreenshot()`. Then edit the resulting screenshot
using an image editing program such as GIMP or Krita so the screenshot
matches the desired color output. From there, load the embedded LUT
texture from the screenshot using `simplepbr.load_sdr_lut()`.